ssh: Make ssh connection more robust just after machine creation
If we ssh in a bit too quickly after the container creation, ssh errors out
with:

  ssh: connect to host 172.17.0.2 port 22: Connection refused

We now filter ssh stderr looking for that message and retry up to 3 times
before giving up.

Because we used -q before to remove the message that is displayed every single
time:

  Warning:Permanently added '172.17.0.2' (ECDSA) to the list of known hosts

we now also filter it out.